Got a new solution in but still not the one I needed for this puzzle
Looking for the first max number and based on that set in the rest of
the algorithm. It still seem to work for the example input, but not for
the full input.
I feel like im getting close. Maybe I should have a look at it again.
Refactor it a bit. But for now I will leave this one.